Moved user-mode events package back into tracee package for the time being
